Add ability to revert all buffers.
This change adds a trivial `anupamk:utils/revert-all-buffers' defun,
which allows to revert all open buffers.

This is quite useful when we change git branches and we want the same
set of files to be immediately available under the aegis of the new
repo f.e. when doing multiple code reivews, working on different set
of changes etc.