emacs: override Emacs keybindings
Depending on major mode vagaries, availability of user specified key
overrides within Emacs via the canonical (define-key ...), tends to be
pretty much a hit-or-miss affair.

With this change we define our own custom keymap, and add them to the
'emulation-mode-map-alists' which takes precedence over everything
else. This ensures that our chosen keybindings are always available.