Fixup for face-attributes and other minor changes
The look-n-feel of an Emacs frame is customized via the
'anupamk:utils/sanitize-frame-look' function. This function is called
during startup as well as after a new frame is created.

This change fixes issues with the usage of 'set-face-attribute' which
was incorrectly setting up the ':background' (and similar attributes
to 'nil).

A minor post-initialization message is also now displayed. This is
cute and useful as well (to ensure that startup configuration was
processed in entirety).