Slightly optimized configuration attribute loading
My per-host emacs configuration is accesses via an 'attribute
table'. During startup, depending on the host where emacs session is
created, appropriate attribute table is loaded.

The expectation was that each host has it's own attribute table, which
turns out to be too much. I don't tend to do *that* much
customization :)

It is so much better to have a default attribute table that is
returned for environments where explicit attribute definitions are not
available.

Which is exactly what this change does.