emacs updated display-buffer-alist configuration
This is a simplification of the notorious display-buffer-alist
configuration. The fundamental difference being that all the
temporary/transient buffers are spawned within the same slot.

As an example, both `man` and `magit` buffers are spawned on the right
side of the Emacs window. Since both of them use `slot-0`, `magit` and
`man` buffers reuse the same window. This is more deterministic
behavior, and one that works for /me/, let see...