Must better display-buffer-alist configuration
This change sets up much better / robust display-buffer-alist
configuration. Specifically, most of the transient buffers that are
created are now not displayed in 'side-windows' and if / when they
are, those windows are not marked 'dedicated'.

This allows for much more smoother overall interaction within
Emacs. Lets just hope that this setup proves to be more long-lasting
and extensible.