Remove usage of 'custom-set-variables' from our configuration
With 'custom-set-variables' a customization file is created, and
populated with appropriate values. In our case, this file is called
'anupam-emacs-custom.el'.

Now, when a variable is edited, and new configuration is generated,
the modified values are not populated once again i.e. the
aforementioned file remains the same as before.

This causes confusion to no end. Fix that.