Add a package and delete a package.
With this change, the 'which-key' package that was disabled for quite
a while now, is now removed.

The 'ultra-scroll' package, which gives quite a remarkable boost and
precision to scrolling within Emacs buffers, is now added.

On the host 'soma' we now use modus-operandi theme, or the modified
version of that theme by default.