Lots of minor improvements all over the place
There is something for everybody :o) ! At a high level though moving
to lspce continues, for more hosts.

On the same note, it seems that ccls is a much better lsp-server than
clangd ! clangd is just too profligate with its resource usage,
spawning multiple number of threads just all over the place. It seems
that it is doing so 1 for every buffer (at the very least !).

Minor gc and other odds and ends improvements.